+++ src/Graphics/GPUMesh.h
@@ -11,14 +11,14 @@ static const auto NULL_MESH_ID = std::numeric_limits<std::size_t>::max();
 
 struct GPUMesh {
     wgpu::Buffer indexBuffer;
-    std::size_t indexBufferSize{0};
+    std::uint32_t indexBufferSize{0};
     wgpu::Buffer vertexBuffer;
 
     MaterialId materialId{NULL_MATERIAL_ID};
 
     struct AttribProps {
-        std::size_t offset;
-        std::size_t size;
+        std::uint64_t offset;
+        std::uint64_t size;
     };
     std::vector<AttribProps> attribs;
